CI using Appveyor and fixes for windows mingw and MSVC build. Executing unit tests on appveyor disabled due to timeout limit.

+++ modules/elm/core/test/signal_unittests.cpp
@@ -36,7 +36,7 @@ protected:
         in.push_back("foo");
         in.push_back("bar");
 
-        for(uint i=0; i<in.size(); i++) {
+        for(size_t i=0; i<in.size(); i++) {
 
             to_.Append(in[i], Mat1f());
         }
